% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{add.event}
\alias{add.event}
\title{Add an event to the timeline (for adding to plots etc)}
\usage{
add.event(str, date = format(Sys.Date(), "\%Y-\%m-\%d"), geoid = "UK")
}
\arguments{
\item{str}{Short string describing event}

\item{date}{Date on which event occurred, yyyy-mm-dd; default is system date}

\item{geoid}{Country in which event is relevant; default is UK}
}
\description{
Add key event to timeline .csv; these can be plotted against case & mortality trajectories for attribution
}
