% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_tools.R
\name{add_map_stats}
\alias{add_map_stats}
\title{cas/morts/gueris pour 100k de la population.}
\usage{
add_map_stats(spatial_df)
}
\arguments{
\item{spatial_df}{Data frame enrichi des valeurs de population par pays.}
}
\value{
Data.frame.
}
\description{
Ces stats serviront aussi au font de carte.
}
\seealso{
Other covid stats: 
\code{\link{add_stats2}()},
\code{\link{add_stats}()},
\code{\link{do_gt_table}()},
\code{\link{global_stats}()}
}
\concept{covid stats}
