% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_tools.R
\name{add_stats}
\alias{add_stats}
\title{Ajouts d'indicateurs a date.}
\usage{
add_stats(global_data)
}
\arguments{
\item{global_data}{Data frame tidy global (cas/morts/gueris).}
}
\value{
Data.frame au format tidy.
}
\description{
Les nouveaux cas, deces, guerisons du jour.
}
\seealso{
Other covid stats: 
\code{\link{add_map_stats}()},
\code{\link{add_stats2}()},
\code{\link{global_stats}()}
}
\concept{covid stats}
