% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_tools.R
\name{add_stats2}
\alias{add_stats2}
\title{Stats recalculées.}
\usage{
add_stats2(global_data)
}
\arguments{
\item{global_data}{Data frame des donnees globales (cas/morts/nouveaux cas).}
}
\value{
Data.frame.
}
\description{
Re comptage des cas comme nouveaux cas.
}
\seealso{
Other covid stats: 
\code{\link{add_map_stats}()},
\code{\link{add_stats}()},
\code{\link{global_stats}()}
}
\concept{covid stats}
