% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_tools.R
\name{add_stats_global}
\alias{add_stats_global}
\title{Stats recalculées pour l'onglet regions/departements.}
\usage{
add_stats_global(df_france_global)
}
\arguments{
\item{df_france_global}{Data frame des donnees globales (cas/morts/gueris).}
}
\value{
Data.frame.
}
\description{
Re comptage des cas (cumul) comme nouveaux cas (à date) pour la France au global.
}
\seealso{
Other covid stats: 
\code{\link{add_map_stats}()},
\code{\link{add_stats2}()},
\code{\link{add_stats_reg}()},
\code{\link{add_stats}()},
\code{\link{do_gt_table}()},
\code{\link{global_stats}()}
}
\concept{covid stats}
