% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_tested_detected_admitted.R
\name{ae_covid_tested_detected_admittted}
\alias{ae_covid_tested_detected_admittted}
\title{ae_covid_tested_detected_admitted}
\usage{
ae_covid_tested_detected_admittted(
  data,
  start_date,
  end_date,
  plot_chart = TRUE,
  hospital_name
)
}
\arguments{
\item{hospital_name}{}
}
\value{

}
\description{
ae_covid_tested_detected_admitted
}
