% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tools.R
\name{base_map_global}
\alias{base_map_global}
\title{Map covid-19 monde vierge.}
\usage{
base_map_global(data_day, map_param, pal_param)
}
\arguments{
\item{data_day}{Data frame des donnees a date.}

\item{map_param}{Objet de type "SpatialPolygonsDataFrame".}

\item{pal_param}{Parametre contenant une fonction de palette de couleurs.}
}
\value{
Objet de type leaflet.
}
\description{
Une base de map reutilisable.

Un fond de carte est applique sur l'indicateur morts/population/100000.
}
\seealso{
\code{\link[=leaflet]{leaflet()}}.
}
\concept{covid map}
