% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c19.plot}
\alias{c19.plot}
\title{Plot reported cases & deaths for a given country}
\usage{
c19.plot(
  ccd = "UK",
  lth = 10,
  ds = 5,
  c.ymx,
  xmin = as.Date("2020-03-01"),
  main = paste0("Reported cases & deaths in ", ccd)
)
}
\arguments{
\item{ccd}{Country code (used to match with geoid in ECDC data)}

\item{lth}{Threshold for adding lines from points to spline. Default is 10.}

\item{ds}{Scaling factor to use when plotting deaths. Default is 5.}

\item{c.ymx}{(optional) maximum value of cases to plot}

\item{xmin}{Date to plot from. Default is as.Date("2020-03-01").}

\item{main}{Main title for plot}
}
\description{
Plot reported cases & deaths for a given country
}
