% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c19.trajectory}
\alias{c19.trajectory}
\title{Trajectory of recent vs total cases for a single country}
\usage{
c19.trajectory(
  ccd = "UK",
  add = F,
  deaths = F,
  flip = F,
  cex = 0.8,
  col = transp("black"),
  ...
)
}
\arguments{
\item{ccd}{Country code to plot. Default is UK}

\item{add}{Boolean: add to existing plot? Default is F (draw new plot)}

\item{deaths}{Boolean: plot deaths instead of cases? Default is F (plot cases)}

\item{flip}{Boolean: flip axes? Default is F (X = total cases, Y = last week)}

\item{...}{Further graphical parameters to be passed to plotting function}
}
\description{
Trajectory of recent vs total cases for a single country
}
