% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{covid}
\alias{covid}
\title{COVID-19 time series}
\usage{
covid(country_region = "Greece")
}
\arguments{
\item{country_region}{country}
}
\description{
COVID-19 time series
}
\examples{
covid(country_region="Germany")
covid(country_region="Greece")
}
\keyword{covid}
