% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_pneumonia_age_gender.R
\name{covid_infected_pneumonia}
\alias{covid_infected_pneumonia}
\title{covid_infected_pneumonia}
\usage{
covid_infected_pneumonia(
  data,
  start_date,
  end_date,
  plot_chart = TRUE,
  hospital_name
)
}
\arguments{
\item{hospital_name}{}
}
\value{

}
\description{
covid_infected_pneumonia
}
