% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_tested_detected_ethnicity.R,
%   R/covid_tst_dtct_adm_died_ethn.R
\name{covid_tst_dtct_adm_died_ethn}
\alias{covid_tst_dtct_adm_died_ethn}
\title{covid_tested_infected_ethnicity}
\usage{
covid_tst_dtct_adm_died_ethn(
  data,
  start_date,
  end_date,
  plot_chart = TRUE,
  hospital_name
)

covid_tst_dtct_adm_died_ethn(
  data,
  start_date,
  end_date,
  plot_chart = TRUE,
  hospital_name
)
}
\arguments{
\item{hospital_name}{}
}
\value{



}
\description{
covid_tested_infected_ethnicity

covid_tested_infected_ethnicity
}
