% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_tools.R
\name{cumul_new_plotly}
\alias{cumul_new_plotly}
\title{Graphique cumul (nouveaux cas) avec plotly.}
\usage{
cumul_new_plotly(df, var_count, col)
}
\arguments{
\item{df}{Data frame pour cuml global/pays/continent.}

\item{var_count}{Variable du data frame correspondant a cas/mors.}

\item{col}{Codes couleurs pays ("#cc4c02", "#d84cff", \code{...})}
}
\value{
plotly object.
}
\description{
Evolution des nouveaux cas/mors \code{...}.

Affichage sous forme de barres geom_bar().
}
\seealso{
\code{\link[=ggplotly]{ggplotly()}} pour plus de détails.

Other plotly tools: 
\code{\link{cumul_plotly_log}()},
\code{\link{cumul_plotly}()}
}
\concept{plotly tools}
