% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_tools.R
\name{cumul_plotly_log}
\alias{cumul_plotly_log}
\title{Graphique cumul log10 avec plotly.}
\usage{
cumul_plotly_log(df, var_count, ylabel, col)
}
\arguments{
\item{df}{Data frame pour cuml global/pays/continent.}

\item{var_count}{Variable du data frame correspondant a cas/mors.}

\item{ylabel}{Label axe des y.}

\item{col}{Codes couleurs pays ("#cc4c02", "#d84cff", \code{...})}
}
\value{
plotly object.
}
\description{
Evolution des nouveaux cas/mors \code{...}.
}
\seealso{
\code{\link[=ggplotly]{ggplotly()}} pour plus de détails.

Other plotly tools: 
\code{\link{cumul_new_plotly}()},
\code{\link{cumul_plotly}()}
}
\concept{plotly tools}
