% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{cumulative_plot}
\alias{cumulative_plot}
\title{Graphique des cas cumules jusqu'a date.}
\usage{
cumulative_plot(df_aggregated, plot_date, col)
}
\arguments{
\item{df_aggregated}{Data frame agrege des donnees globales (cas/morts/nouveaux cas).}

\item{plot_date}{Date en caracteres ("2020-01-25").}

\item{col}{Code couleur ("#cc4c02")}
}
\value{
ggplot2 object.
}
\description{
Evolution du comptage des cas.
}
\seealso{
Other plot tools: 
\code{\link{new_cases_plot}()}
}
\concept{plot tools}
