% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{daily.cases}
\alias{daily.cases}
\title{Plot daily cases}
\usage{
daily.cases(
  incl = c("UK", "CN", "JP", "KR", "IT", "ES", "FR", "DE", "US"),
  ccols = c("black", rbow(length(incl) - 1)),
  show.China = F,
  smooth = T,
  add.lockdowns = T,
  ymx = NA
)
}
\arguments{
\item{incl}{Vector of country codes to include. Default is c("UK", "CN", "JP", "KR", "IT", "ES", "FR", "DE", "US").}

\item{ccols}{Vector of colours to use for each country. Default is black for first country and rainbow palette for the remainder.}

\item{show.China}{Boolean: show full China trajectory or truncate to European days since 100th confirmed case? Default is F (truncated).}

\item{smooth}{Boolean: apply a spline smoother? Default is T.}

\item{add.lockdowns}{Boolean: add points showing when lockdowns were introduced? Default is T.}
}
\description{
Plot daily cases
}
