% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{daily.deaths}
\alias{daily.deaths}
\title{Plot daily deaths}
\usage{
daily.deaths(
  incl = c("UK", "CN", "JP", "KR", "IT", "ES", "FR", "DE", "US"),
  ccols = c("black", rbow(length(incl) - 1)),
  offset.dates = NA,
  show.China = F,
  smooth = "o",
  ymx = NA,
  leg.ncols = 2,
  main
)
}
\arguments{
\item{incl}{Vector of country codes to include. Default is c("UK", "CN", "JP", "KR", "IT", "ES", "FR", "DE", "US").}

\item{ccols}{Vector of colours to use for each country. Default is black for first country and rainbow palette for the remainder.}

\item{offset.dates}{Character: show days since 100th case ("c"), days since 10th death ("d"), or date? Default is date.}

\item{show.China}{Boolean: show full China trajectory or truncate? Default is F (truncated).}

\item{smooth}{Boolean: apply a spline smoother? Default is T.}

\item{ymx}{Maximum value to show on y-axis}

\item{leg.ncols}{Number of columns in legend. Default is 2.}

\item{add.lockdowns}{Boolean: add points showing when lockdowns were introduced? Default is F.}
}
\description{
Plot daily deaths
}
