% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_covid}
\alias{data_covid}
\title{Import depuis repos Github.}
\usage{
data_covid(data_github)
}
\arguments{
\item{data_github}{Lien html vers fichier .csv (repos Github).}
}
\value{
Objet de type data.table data.frame.
}
\description{
Donnés COVID au format ".csv".
}
\seealso{
\href{https://github.com/CSSEGISandData/COVID-19}{sources} les sources.

Other data tools: 
\code{\link{data_recod}()},
\code{\link{make_tidy_df}()}
}
\concept{data tools}
