% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{data_source}
\alias{data_source}
\title{Display data source}
\usage{
data_source(id)
}
\arguments{
\item{id}{A character vector of length 1.}
}
\value{
A HTML element.
}
\description{
Display data source
}
\examples{
\dontrun{
data_source("usa")
}
}
