% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{deaths.since.d10}
\alias{deaths.since.d10}
\title{Replicate Chartr plot showing number of deaths vs time since hundredth case}
\usage{
deaths.since.d10(
  incl = c("UK", "CN", "JP", "KR", "IT", "ES", "FR", "DE", "US"),
  ccols = c("black", rbow(length(incl) - 1)),
  doublings = 3,
  show.China = F,
  log = F,
  main = "Selected mortality trajectories"
)
}
\arguments{
\item{incl}{Vector of country codes to include. Default is c("UK", "CN", "JP", "KR", "IT", "ES", "FR", "DE", "US").}

\item{ccols}{Vector of colours to use for each country. Default is black for first country and rainbow palette for the remainder.}

\item{doublings}{Add trajectories for cases doubling every n days.}

\item{show.China}{Boolean: show full China trajectory or truncate to European days since 100th confirmed case? Default is F (truncated).}

\item{log}{Boolean: plot on log scale? Default is F (show on linear scale)}
}
\description{
Replicate Chartr plot showing number of deaths vs time since hundredth case
}
