% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_tools.R
\name{do_gt_table}
\alias{do_gt_table}
\title{Tableau dynamique}
\usage{
do_gt_table(df, niveau)
}
\arguments{
\item{df}{Data frame pour onglet "pays".}

\item{niveau}{Ex : "pays", "continent"}
}
\value{
Data.frame.
}
\description{
Tableau avec presentation \code{\link[=gt]{gt()}}
}
\seealso{
Other covid stats:
\code{\link{add_map_stats}()},
\code{\link{add_stats2}()},
\code{\link{add_stats}()},
\code{\link{global_stats}()}
}
\concept{covid stats}
