% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{doubling}
\alias{doubling}
\title{Add line to plot showing trajectory of a given doubling rate}
\usage{
doubling(nd = 3, label.at = 25, baseline = 100)
}
\arguments{
\item{nd}{Number of days for cases to double}
}
\description{
Add line to plot showing trajectory of a given doubling rate
}
