% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_bar_chart}
\alias{draw_bar_chart}
\title{Draw area bar chart}
\usage{
draw_bar_chart(df, value_var)
}
\arguments{
\item{df}{A data frame with columns: publish_date, population,
prefecture, state or area,
concept, value, value_ma, value_per1m, value_ma_per1m.}

\item{value_var}{A symbol: value, value_ma, value_per1m or value_ma_per1m.}
}
\value{
A plot.
}
\description{
Draw area bar chart
}
\examples{
\dontrun{
draw_bar_chart(df, value_ma)
}
}
