% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_line_chart}
\alias{draw_line_chart}
\title{Draw line chart}
\usage{
draw_line_chart(df, var_str)
}
\arguments{
\item{df}{A data frame with columns: publish_date, concept, value, var.}

\item{var_str}{A character vector of length 1.}
}
\value{
A plot.
}
\description{
Draw line chart
}
\examples{
\dontrun{
draw_line_chart(df, var_str)
}
}
