% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_map_japan}
\alias{draw_map_japan}
\title{Draw Japan prefecture map}
\usage{
draw_map_japan(df, value_var)
}
\arguments{
\item{df}{A data frame with columns: publish_date, concept, value, code.}

\item{value_var}{"value", "value_ma", "value_per1m" or "value_ma_per1m".}
}
\value{
A plot.
}
\description{
Draw Japan prefecture map
}
\examples{
\dontrun{
draw_map_japan(df, "value_ma")
}
}
