% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_map_japan_simple}
\alias{draw_map_japan_simple}
\title{Draw Japan prefecture map}
\usage{
draw_map_japan_simple(df, var_str)
}
\arguments{
\item{df}{A data frame with columns: publish_date, concept, value, code.}

\item{var_str}{A character vector of length 1.}
}
\value{
A plot.
}
\description{
Draw Japan prefecture map
}
\examples{
\dontrun{
draw_map_japan_simple(df, "value_ma")
}
}
