% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_map_usa_simple}
\alias{draw_map_usa_simple}
\title{Draw USA state map}
\usage{
draw_map_usa_simple(df, var_str)
}
\arguments{
\item{df}{A data frame with columns: publish_date, concept, value, state.}

\item{var_str}{A character vector of length 1.}
}
\value{
A plot.
}
\description{
Draw USA state map
}
\examples{
\dontrun{
draw_map_usa_simple(df, "value")
}
}
