% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{draw_mapUI}
\alias{draw_mapUI}
\title{Draw map module UI}
\usage{
draw_mapUI(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}
}
\value{
A module UI.
}
\description{
Draw map module UI
}
\examples{
\dontrun{
draw_mapUI("world_map", world$area_df)
}
}
