% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMap.R
\name{drawMapServer}
\alias{drawMapServer}
\title{Draw map module Server}
\usage{
drawMapServer(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame. Columns: publish_date, concept and value_map.}
}
\value{
A module server.
}
\description{
Draw map module Server
}
\examples{
\dontrun{
drawMapServer("usa_map", usa$area_df)
}
}
