% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMap.R
\name{drawMapUI}
\alias{drawMapUI}
\title{Draw map module UI}
\usage{
drawMapUI(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}
}
\value{
A module UI.
}
\description{
Draw map module UI
}
\examples{
\dontrun{
drawMapUI("world_map", world$area_df)
}
}
