% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMapWk.R
\name{drawMapWkServer}
\alias{drawMapWkServer}
\title{Draw map weekly module Server}
\usage{
drawMapWkServer(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame. Columns: publish_date, concept and value_map.}
}
\value{
A module server.
}
\description{
Draw map weekly module Server
}
\examples{
\dontrun{
drawMapWkServer("japan_map", japan_wk$area_df)
}
}
