% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMapWk.R
\name{drawMapWkUI}
\alias{drawMapWkUI}
\title{Draw map weekly module UI}
\usage{
drawMapWkUI(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}
}
\value{
A module UI.
}
\description{
Draw map weekly module UI
}
\examples{
\dontrun{
drawMapWkUI("japan_map", japan_wk$area_df)
}
}
