% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECDC-daily.R
\name{ecdc.daily}
\alias{ecdc.daily}
\title{Refresh ECDC data & plots}
\usage{
ecdc.daily(
  countries = c("UK", "CN", "KR", "IT", "ES", "US"),
  ccols = c("black", "forestgreen", "blue3", "dodgerblue", "darkgoldenrod2", "red")
)
}
\arguments{
\item{ccols}{Vector of colours to use for countries in incl}

\item{incl}{Vector of country codes to include}
}
\description{
Download latest data from ECDC & produce pdf of standard plots
}
