% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction rate estimation.R
\name{estimateR}
\alias{estimateR}
\title{Estimate reproduction rate for a time series of incidence}
\usage{
estimateR(
  incidence = ita \%>\% filter(region == "Lombardy") \%>\% pull(pos.total_it),
  total = TRUE,
  l.interval = 2,
  warnings = FALSE,
  plot = FALSE,
  mean.si = 4.7,
  sd.si = 2.9
)
}
\arguments{
\item{sd.si}{}
}
\value{

}
\description{
Standards are based on paper Serial interval of novel coronavirus (COVID-19) infections
"while the mean and standard deviation (SD) of the serial interval were estimated at 4.7 days (95% CrI: 3.7, 6.0) and 2.9 days (95% CrI: 1.9, 4.9), respectively"
}
\examples{
estimateR(plot=TRUE)
estimateR(plot=TRUE,mean.si=7.5,sd.si=3.4,l.interval=2)
}
