% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_covid}
\alias{etl_extract.etl_covid}
\title{etl_extract.etl_covid}
\usage{
\method{etl_extract}{etl_covid}(obj, month, year, ...)
}
\arguments{
\item{obj}{an \code{\link[etl]{etl}} object}

\item{month}{fill in}

\item{year}{fill in}

\item{...}{arguments passed to methods}
}
\description{
etl_extract.etl_covid
}
\details{
This function obtains the daily reports of COVID-19 from the John Hopkins CSSE repo.
Users can specify the month and year they would like to obtain data for. However,
all data is pulled when no arguments are specified. The downloaded dataframes
are saved in the directory the user specified.
}
