% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_load.R
\name{etl_load.etl_covid}
\alias{etl_load.etl_covid}
\title{Load Covid-19 Data into PostgreSQL Database}
\usage{
\method{etl_load}{etl_covid}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link[etl]{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
Load Covid-19 Data into PostgreSQL Database
}
