% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{events}
\alias{events}
\title{Load timeline (for adding to plots etc)}
\usage{
events(geoid = NA)
}
\arguments{
\item{geoid}{Vector of country IDs to include. Default is NA, which includes all events}
}
\description{
Load timeline (for adding to plots etc)
}
