% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{events.with.stats}
\alias{events.with.stats}
\title{Load timeline with cumulative cases}
\usage{
events.with.stats(incl = NA)
}
\arguments{
\item{incl}{Vector of country codes to include}
}
\description{
Load timeline with cumulative cases
}
