% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm helpers.R
\name{expand_formula}
\alias{expand_formula}
\title{Expands formulas for plm}
\usage{
expand_formula(form = "pos.growth ~.", df)
}
\arguments{
\item{form}{Formula where ~. is substituted}

\item{dataframe}{for colnames}
}
\value{

}
\description{
Expands formulas for plm
}
