% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_area_var.R
\name{extract_area_var}
\alias{extract_area_var}
\title{Extract area variable symbol from a data frame}
\usage{
extract_area_var(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A symbol.
}
\description{
Extract area variable symbol from a data frame
}
\examples{
\dontrun{
extract_area_var(world$region_df)
}
}
