% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags helpers.R
\name{extract_mean}
\alias{extract_mean}
\title{Extract mean from MCMC summary object}
\usage{
extract_mean(res, var = "R\\\\[", without = NULL)
}
\arguments{
\item{res}{mcmc object}

\item{var}{string descirbing variables to extract}
}
\value{
data.frame with t,country, and var
}
\description{
Extract mean from MCMC summary object
}
