% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_str.R
\name{extract_var_str}
\alias{extract_var_str}
\title{Extract area variable from a data frame}
\usage{
extract_var_str(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A character vector of length 1.
}
\description{
Extract area variable from a data frame
}
\examples{
\dontrun{
extract_var_str(world$region_df)
}
}
