% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm helpers.R
\name{factorize}
\alias{factorize}
\title{Factorize continuous variables by quantiles of ecdf}
\usage{
factorize(x, length = 5)
}
\arguments{
\item{x}{variable to be factorized}

\item{length}{number of factors}
}
\value{
vector
}
\description{
Factorize continuous variables by quantiles of ecdf
}
\examples{
unique(factorize(all$tMax,7))
}
