% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_forecast.R
\name{forecast}
\alias{forecast}
\title{Application of Prophet methods for covid-19 total case prediction}
\usage{
forecast(country, data, numPred)
}
\arguments{
\item{country}{A character input corresponding to the country modeled}

\item{data}{A data.frame input with country name, total cases reported, date, and total population. 
This can be the dataset for all countries or an individual country}

\item{numPred}{A numeric input which indicates how many days in the futute predictions should be made.}
}
\value{
A graph displaying the model, previous cases, and predictions and a dataset which adds predicted 
counts onto data input to the function
}
\description{
This function creates a prophet model  and predicts future cases for individual countries in the
John's Hopkins COVID-19 database. The dataset must also include a capacity population in order to
properly run the logistic modeling. The function prints a graph of previously reported cases, the 
model determined by prophet, and future predictions as well as a data.frame with predicted cases 
included.
}
\examples{

forecast("US", dat3, 7)
forecast("Korea, S", dat3, 7)

}
