% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get.ecdc}
\alias{get.ecdc}
\title{Download latest ECDC data}
\usage{
get.ecdc(dl.dt = format(Sys.Date(), "\%Y-\%m-\%d"))
}
\description{
Download latest Covid-19 infection data from ECDC website and prep for plotting etc
}
