% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_var.R
\name{get_value_var}
\alias{get_value_var}
\title{Generate value variable symbol}
\usage{
get_value_var(ma, per1m)
}
\arguments{
\item{ma}{"Yes" or "No".}

\item{per1m}{"Yes" or "No".}
}
\value{
A symbol.
}
\description{
Generate value variable symbol
}
\examples{
\dontrun{
get_value_var("Yes", "No")
}
}
