% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRF.R
\name{getRF}
\alias{getRF}
\title{Generation of a graph for any country's number of new COVID-19 cases overlayed with a Random Forest Model}
\usage{
getRF(Country_Name, Pred_Day = 8)
}
\arguments{
\item{Country_Name}{A character input corresponding to the country being graphed}

\item{Pred_Day}{A specified number of Days past from April 3rd the user wants to predict. Default is 8 days.}
}
\value{
A graph displaying the number of new cases for specified country predicted through a random forest model.
}
\description{
This function creates a graph through a random forest prediction framework for any 
user-specified country that displays number of new cases
of COVID-19 from their baseline day of 50 cases confirmed until April 3, 2020,
with further predictions specified by the user.
}
\examples{

}
