% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_tools.R
\name{global_stats}
\alias{global_stats}
\title{Stats monde.}
\usage{
global_stats(global_df)
}
\arguments{
\item{global_df}{Data frame des donnees globales (cas/morts/nouveaux cas).}
}
\value{
Data.frame aggrege par date.
}
\description{
Aggregation des donnees.
}
\seealso{
Other covid stats: 
\code{\link{add_map_stats}()},
\code{\link{add_stats}()}
}
\concept{covid stats}
