% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{japan_wk}
\alias{japan_wk}
\title{Japan Covid-19 situation, weekly}
\format{
A list.
\describe{
\item{area_df}{

A tibble.
\describe{
\item{publish_date}{2020-01-16}
\item{code}{"01"}
\item{prefecture}{"Hokkaido"}
\item{concept}{"new_conf_wk"}
\item{value}{0}
\item{value_per_hospital}{0}
}
}
\item{area_menu}{

A character vector of length 47.
\describe{
\item{1}{"Hokkaido"}
\item{2}{"Aomori"}
\item{3}{"Iwate"}
\item{4}{"Miyagi"}
\item{5}{"Akita"}
}
}
}
}
\source{
https://www.mhlw.go.jp/stf/seisakunitsuite/bunya/0000121431_00432.html
}
\usage{
japan_wk
}
\description{
Japan Covid-19 situation, weekly
}
\keyword{datasets}
