% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data manipulations.R
\name{keep_units_with}
\alias{keep_units_with}
\title{Drops any units (countries) that do not have measures imposed}
\usage{
keep_units_with(df = world, measures = world.measures)
}
\arguments{
\item{df}{data frame to be subsetted}

\item{measures}{vector of measures to be considered (keeps any measures that submatch string)}
}
\value{
pdata.frame
}
\description{
Drops any units (countries) that do not have measures imposed
}
\examples{
keep_units_with()
}
