% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lag}
\alias{lag}
\title{Calculate lag between two countries}
\usage{
lag(c1, c2, diffs = F)
}
\arguments{
\item{c1}{Country code of first country to compare}

\item{c2}{Country code of second country to compare}

\item{diffs}{Boolen: return vector of differences on lags -50:50? Default is F (return lag).}
}
\description{
Calculate lag between two countries
}
