% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lagged.cases}
\alias{lagged.cases}
\title{Plot cumulative cases aligned with another country's trajectory}
\usage{
lagged.cases(
  c1,
  c2,
  log = F,
  show.original = F,
  show.lockdowns = F,
  ccols = rbow(length(c2) + 1)[-1]
)
}
\arguments{
\item{c1}{Country to use as baseline}

\item{c2}{Vector of countries to compare}

\item{log}{Boolean: use log scale? Default is F}

\item{show.original}{Boolean: show actual dates as well as offset dates? Default is F}

\item{ccols}{Vector of colours to use for plotting countries in c2. Default is rainbow.}
}
\description{
Plot cumulative cases aligned with another country's trajectory
}
