% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{lockdowns}
\alias{lockdowns}
\title{Extract lockdown events with associated data}
\usage{
lockdowns(ccd)
}
\arguments{
\item{ccd}{Vector of country codes to extract lockdowns for}
}
\description{
Extract lockdown events with associated data
}
