% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lw.cc.global}
\alias{lw.cc.global}
\title{Plot weekly cases vs cumulative}
\usage{
lw.cc.global(
  incl = c("UK", "CN", "JP", "KR", "IT", "ES", "FR", "DE", "US"),
  ccols = c("black", rbow(length(incl) - 1)),
  log = T,
  grid = F,
  deaths = T,
  main = NA
)
}
\arguments{
\item{incl}{Vector of country codes to include. Default is c("UK", "CN", "JP", "KR", "IT", "ES", "FR", "DE", "US").}

\item{ccols}{Vector of colours to use for each country. Default is black for first country and rainbow palette for the remainder.}

\item{log}{Boolean: plot on log scale? Default is T (show on log10 scale)}

\item{grid}{Boolean: add a grid? Default is F.}

\item{main}{Plot title: default is "Trajectory of confirmed cases"}
}
\description{
Plot weekly cases vs cumulative
}
