% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flow_groups.R
\name{make_flow_groups}
\alias{make_flow_groups}
\title{make_flow_groups}
\usage{
make_flow_groups(spell_data)
}
\arguments{
\item{spell_data}{Hospital flow data with columns for
ed non-admission (boolean);
directorate (e.g. "Medical", "Surgical")}
}
\value{
spell_data with additional column for flow group
}
\description{
make_flow_groups
}
