% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{make_tidy_df}
\alias{make_tidy_df}
\title{Structuration des donnees.}
\usage{
make_tidy_df(data)
}
\arguments{
\item{data}{Objet de type data frame (time series cases/deaths/recovered).}
}
\value{
Data.frame au format tidy.
}
\description{
Mise au format "tidy".

L'objectif est de faciliter les jointures.
}
\seealso{
Other data tools: 
\code{\link{data_covid}()},
\code{\link{data_recod_dom}()},
\code{\link{data_recod}()}
}
\concept{data tools}
