% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{map.totals}
\alias{map.totals}
\title{Draw map & colour according to data from mappable.summ()}
\usage{
map.totals(varb = "tcases", cols = rbow(11), intvl, ...)
}
\arguments{
\item{varb}{Variable to use to colour map. Default is tcases.}

\item{cols}{Vector of colours to use in plotting. Default is rbow(11).}

\item{intvl}{Vector of intervals to be used to cut variable. Must be one less than length(cols).}

\item{...}{Any other optional arguments to be passed to map("world", ...)}
}
\description{
Draw map & colour according to data from mappable.summ()
}
