% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mappable.summ}
\alias{mappable.summ}
\title{Match summary to countries used in mapping}
\usage{
mappable.summ(...)
}
\arguments{
\item{countries}{Vector of country names to match, obtained from map("world", plot = F)$names}
}
\description{
Match summary to countries used in mapping
}
