% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwgrw.R
\name{mwgrw}
\alias{mwgrw}
\title{Estimation of a Generalized Linear Mixed Model of COVID-19 data using a Metropolis-within-Gibbs Random Walk}
\usage{
mwgrw(
  beta = NA,
  Sigma_gamma = NA,
  M = 1000,
  Z = c(1, 2),
  control = list(tol = 10^-5, maxit = 1000),
  verbose = TRUE
)
}
\arguments{
\item{beta}{An intial vector of length 6 to be the starting values of beta for the MCEM algorithm}

\item{Sigma_gamma}{An intial matrix of size qxq, where q is the dimension of the random effects. Default random effects would require a 2x2 matrix.}

\item{M}{Chain length for each country in the MCEM algorithmn. Defaults to 1000.}

\item{Z}{A vector that contains the subset of columns of X to be random effects. Defaults to intercept and time only.}

\item{control}{A list of control parameters for the MCEM algorithm. tol is the convergence tolerance, and maxit is the maximum number of iterations}

\item{verbose}{A boolean indicating if output should be printed at each iteration}
}
\value{
A list containing: beta - final estimates of beta, Sigma_gamma - final estimates of Sigma_gamma, gamma - chains from the final iteration of the EM algorithm, and time - time the MCEM algorithm took to run.
}
\description{
This function computes the estimates of a Poisson generalized linear mixed model
with a log link for COVID-19 data from the Johns Hopkins CSSE. Other data included
comes from the World Bank and the Global Health Secuiry Index. The response is the
number of new cases, while the predictors include an intercept, time, time^2, GHS score,
percentage of population over age 65, and percentage of population living in urban areas.
Random effects can be adjust but is defaulted to intercept and time.
}
