% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwgrwMSE.R
\name{mwgrwMSE}
\alias{mwgrwMSE}
\title{Calculation of MSE}
\usage{
mwgrwMSE(country_name = NULL, glmer_MSE = FALSE)
}
\arguments{
\item{country_name}{A character input corresponding to the country, if left empty the function will calculate the MSE for all countries together}

\item{glmer_MSE}{A boolean determining if the MSE should be calculated for the MCEM GLMM or the glmer function}
}
\value{
The MSE given the parameters specified
}
\description{
This function calculates the MSE for a specific country or every country based on their
predictions for the 8 days following April, 03, 2020.
}
\examples{

mwgrwMSE()
mwgrwMSE(country_name='US')

}
