% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm helpers.R
\name{mypanel}
\alias{mypanel}
\title{make to pdata.frame with index i and t}
\usage{
mypanel(df)
}
\arguments{
\item{df}{data frame to be coverted}
}
\value{
pdata.frame
}
\description{
Use after dplyr commands on data frames
}
