% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_file.R
\name{name_file}
\alias{name_file}
\title{Define download file name}
\usage{
name_file(select_concept, ma, per1m)
}
\arguments{
\item{select_concept}{A character vector of length 1.}

\item{ma}{A character vector of length 1.}

\item{per1m}{A character vector of length 1.}
}
\value{
A character vector of length 1.
}
\description{
Define download file name
}
\examples{
\dontrun{
name_file("new_conf", "Yes", "No")
}
}
