% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{new_cases_plot}
\alias{new_cases_plot}
\title{Graphique des nouveaux cas jusqu'a date.}
\usage{
new_cases_plot(df_aggregated, plot_date, col)
}
\arguments{
\item{df_aggregated}{Data frame agrege des donnees globales (cas/morts/nouveaux cas).}

\item{plot_date}{Date en caracteres ("2020-01-25").}

\item{col}{Code couleur ("#cc4c02")}
}
\value{
ggplot2 object.
}
\description{
Evolution du comptage des nouveaux cas.
}
\seealso{
Other plot tools: 
\code{\link{cumulative_plot}()}
}
\concept{plot tools}
