% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONS.R
\name{occurrences.by.age}
\alias{occurrences.by.age}
\title{Occurrences by age}
\usage{
occurrences.by.age(
  fnm = "./ONS/publishedweek182020.xlsx",
  snm = "Covid-19 - Weekly occurrences",
  cols = "C:BC",
  header.row = "6",
  m.rows = "34:53",
  f.rows = "56:75"
)
}
\arguments{
\item{fnm}{File name to load from. Default is "./ONS/publishedweek152020.xlsx"}

\item{snm}{Sheet name or index. Default is "Covid-19 - Weekly occurrences"}

\item{cols}{Columns to read. Default is "C:BC"}

\item{header.row}{Row containing dates. Default is "6"}

\item{m.rows}{Rows to read for male mortality. Default is "34:53"}

\item{f.rows}{Rows to read for female mortality. Default is "56:75"}
}
\description{
Occurrences by age
}
