% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm helpers.R
\name{panel_reg}
\alias{panel_reg}
\title{Run panel regressios with fixed effects}
\usage{
panel_reg(formula.cur, df)
}
\arguments{
\item{formula.cur}{a valid formular}

\item{df}{a data frame}
}
\value{
output of stargazer
}
\description{
Run panel regressios with fixed effects
}
