% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHE-daily.R
\name{phe.daily}
\alias{phe.daily}
\title{Refresh PHE data & plots}
\usage{
phe.daily(
  highlight = c("Hertfordshire", "Norfolk"),
  hcols = c("black", "orange"),
  archive = F
)
}
\description{
Download latest data from PHE & produce pdf of standard plots
}
