% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covid_cases.R
\name{plot_covid_cases}
\alias{plot_covid_cases}
\title{plot_covid_cases}
\usage{
plot_covid_cases(
  dt_covid,
  country_region,
  province_state,
  category = c("Deaths", "Confirmed", "Recovered"),
  log = FALSE
)
}
\description{
plot_covid_cases
}
