% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prop.cases}
\alias{prop.cases}
\title{Plot cases vs population size}
\usage{
prop.cases(
  incl = c("UK", "US", "CN", "IT", "BE", "BR", "DE", "ES", "FR", "IN"),
  icols = c("magenta3", rbow(length(incl))[-1]),
  deaths = F,
  legend.rows = 5,
  legend = T,
  add.ONS = T
)
}
\arguments{
\item{incl}{Vector of country codes to include}

\item{icols}{Vector of colours to use for each country in incl}

\item{deaths}{Boolean: plot deaths? Default is F (plot cases)}

\item{legend.rows}{Integer: how many rows in legend? Default is 5.}

\item{log}{Boolean: use a log scale? Default is T.}
}
\description{
Plot cases vs population size
}
