% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_show.R
\name{select_showServer}
\alias{select_showServer}
\title{Select show module Server}
\usage{
select_showServer(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}
}
\value{
A module server.
}
\description{
Select show module Server
}
\examples{
\dontrun{
select_showServer("world_region", world$region_df)
}
}
