% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_show.R
\name{select_showUI}
\alias{select_showUI}
\title{Select show module UI}
\usage{
select_showUI(id, df, var_str, a_menu)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}

\item{var_str}{A string of var.}

\item{a_menu}{A named character vector, area menu.}
}
\value{
A module UI.
}
\description{
Select show module UI
}
\examples{
\dontrun{
select_showUI("world_region", world$region_df, "region", world$region_menu)
}
}
