% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectShow.R
\name{selectShowServer}
\alias{selectShowServer}
\title{Select show module Server}
\usage{
selectShowServer(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}
}
\value{
A module server.
}
\description{
Select show module Server
}
\examples{
\dontrun{
selectShowServer("world_region", world$region_df)
}
}
