% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectShow.R
\name{selectShowUI}
\alias{selectShowUI}
\title{Select show module UI}
\usage{
selectShowUI(id, df, a_menu)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}

\item{a_menu}{A named character vector, area menu.}
}
\value{
A module UI.
}
\description{
Select show module UI
}
\examples{
\dontrun{
selectShowUI("world_region", world$region_df, world$region_menu)
}
}
