% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectShowWk.R
\name{selectShowWkServer}
\alias{selectShowWkServer}
\title{Select show weekly module Server}
\usage{
selectShowWkServer(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}
}
\value{
A module server.
}
\description{
Select show weekly module Server
}
\examples{
\dontrun{
selectShowWkServer("japan_wk", japan_wk$area_df)
}
}
