% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectShowWk.R
\name{selectShowWkUI}
\alias{selectShowWkUI}
\title{Select show weekly module UI}
\usage{
selectShowWkUI(id, df, a_menu)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with columns: publish_date, concept, value and var.}

\item{a_menu}{A named character vector, area menu.}
}
\value{
A module UI.
}
\description{
Select show weekly module UI
}
\examples{
\dontrun{
selectShowWkUI("japan_wk", japan_wk$area_df, japan_wk$area_menu)
}
}
