% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time series plots.R
\name{show_countries_of}
\alias{show_countries_of}
\title{Plot timeline of all measures with countries}
\usage{
show_countries_of(
  measure_name = c("University", "School", "Curfew"),
  df = world \%>\% filter(is.na(in_country))
)
}
\arguments{
\item{measure_name}{(also takes vector of names) keeps all measures where substring matches}

\item{df}{data frame}
}
\value{
ggplot
}
\description{
Country/region lables without any measure enacted are droped
}
\examples{
show_countries_of()
}
