% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time series plots.R
\name{show_outcome}
\alias{show_outcome}
\title{show outcome as time series including measures taken}
\usage{
show_outcome(
  select_region = "Lombardy",
  select_country = NULL,
  outcome = "pos.total_it"
)
}
\arguments{
\item{select_region}{region to show}

\item{select_country}{country to show}

\item{outcome}{character variable describing outcome in all data frame}
}
\value{

}
\description{
show outcome as time series including measures taken
}
\examples{
show_outcome()
}
