% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction rate estimation.R
\name{addR}
\alias{addR}
\title{Add R (reproduction rate estimate) to data frame}
\usage{
addR(x, date = NULL, total = TRUE, ...)
}
\arguments{
\item{total}{}
}
\value{

}
\description{
Should be used by group as in example
}
\examples{
res <- ita \%>\%
   group_by(i) \%>\%
   mutate(R = addR(pos.total_it,date,l.interval=7)) \%>\% ungroup()
res \%>\% select(R,pos.total_it,date,region)\%>\%filter(!is.na(R))
}
