% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrygraph.R
\name{countrygraph}
\alias{countrygraph}
\title{Generation of a graph for any country's number of new COVID-19 cases overlayed with Generalized Linear Mixed Model}
\usage{
countrygraph(Country_Name)
}
\arguments{
\item{Country_Name}{A character input corresponding to the country being graphed}
}
\value{
A graph displaying the number of new cases for specified country in black and the GLMM model of new cases in red
}
\description{
This function creates a graph for user specified country that displays number of new cases
of COVID-19 from their baseline day of 50 cases confirmed until April 3, 2020.
Generalized Linear Mixed Model is graphed over the same period to show accuracy of 
model in explaining number of new cases for that country.
}
\examples{

countrygraph(Country_Name='US')
countrygraph(Country_Name='China')




}
