% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrygraph.R
\name{countrygraph}
\alias{countrygraph}
\title{Generation of a graph for any country's number of new COVID-19 cases overlayed with Generalized Linear Mixed Model}
\usage{
countrygraph(
  Country_Name,
  prediction = FALSE,
  Pred_Day = NULL,
  glmer_results = FALSE
)
}
\arguments{
\item{Country_Name}{A character input corresponding to the country being graphed}

\item{prediction}{A boolean input to generate the predictions over the next 8 days}

\item{Pred_Day}{A specified number of days past April 3rd the user wants to predict. Default is 8 days; only used when prediction is set to TRUE}

\item{glmer_results}{A boolean input to display results from glmer in blue}
}
\value{
A graph displaying the number of new cases for specified country in black and the GLMM model of new cases in red, with a dotted extension if prediction is specified as true. GLMER graphed in blue if set to TRUE.
}
\description{
This function creates a graph for user specified country that displays (in black) observed number of new cases
of COVID-19 from their baseline day of 50 cases confirmed until April 3, 2020.
The Poisson Generalized Linear Mixed Model is graphed over the same period (in red) to show accuracy of 
model in explaining number of new cases for that country.
By specifiying glmer_results=TRUE, the glmer results are also graphed over the same period (in blue).
An option to generate the predictions for the next 8 days is present as well. If set to true, predictions will be displayed as
a dotted line extension of the GLMM (red) or glmer (blue) curves.
}
\examples{

countrygraph(Country_Name='US')
countrygraph(Country_Name='US', prediction = TRUE)
countrygraph(Country_Name='US', prediction = TRUE, Pred_Day=50, glmer_results= TRUE)
countrygraph(Country_Name='China')
countrygraph(Country_Name='China', prediction = TRUE)


}
