% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_line_chart}
\alias{draw_line_chart}
\title{Draw line chart}
\usage{
draw_line_chart(df, area_var, value_var)
}
\arguments{
\item{df}{A data frame with columns: publish_date, concept, value, var.}

\item{area_var}{"area", "region", "state" or "prefecture".}

\item{value_var}{"value", "value_ma", "value_per1m" or "value_ma_per1m".}
}
\value{
A plot.
}
\description{
Draw line chart
}
\examples{
\dontrun{
draw_line_chart(df, area_var, value_var)
}
}
