% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{draw_mapServer}
\alias{draw_mapServer}
\title{Draw map module Server}
\usage{
draw_mapServer(id, df)
}
\arguments{
\item{id}{A character vector of length 1.}

\item{df}{A data frame with 3 columns: publish_date, concept and value.}
}
\value{
A module server.
}
\description{
Draw map module Server
}
\examples{
\dontrun{
draw_mapServer("usa_map", usa$ma7_df)
}
}
