% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_covid}
\alias{etl_extract.etl_covid}
\title{Scrape COVID-19 Daily Reports from Github}
\usage{
\method{etl_extract}{etl_covid}(obj, month, year, ...)
}
\arguments{
\item{obj}{an \code{\link[etl]{etl}} object}

\item{month}{numeric vector specifying month(s)}

\item{year}{numeric vector specifying year(s)}

\item{...}{arguments passed to methods}
}
\description{
\code{etl_extract} obtains COVID-19 daily reports uploaded by CSSEGISandData on Github.
Arguments, such as month and year let users obtain data for a specified time period.
When month and year are not specified all data is scraped from the repo. The downloaded
datsets are saved in the folder the user specified or in the temp folder if no
folder was specified.
}
