% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_covid}
\alias{etl_extract.etl_covid}
\title{Scrape COVID-19 Daily Reports from Github}
\usage{
\method{etl_extract}{etl_covid}(obj, month, day, year, ...)
}
\arguments{
\item{obj}{an \code{\link[etl]{etl}} object}

\item{month}{numeric vector specifying month(s)}

\item{day}{numeric vector specifying day(s)}

\item{year}{numeric vector specifying year(s)}

\item{...}{arguments passed to methods}
}
\description{
\code{etl_extract} obtains COVID-19 daily reports uploaded by CSSEGISandData on Github.
Arguments, such as month, day, and year let users obtain data for a specified time period.
When no arguments are specified then all available data is scraped from github. The downloaded
datsets are then saved in the folder that the user specified or in the temp folder if no
folder was specified.
}
\examples{

covid_dat <- etl("covid")

covid_dat \%>\% etl_transform()

covid_dat \%>\% etl_transform(month = 4, year = 2020)

covid_dat \%>\% etl_transform(month = c(3,4), year = 2020)

}
