% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_transform.R
\name{etl_transform.etl_covid}
\alias{etl_transform.etl_covid}
\title{etl_transform.etl_covid}
\usage{
\method{etl_extract}{etl_covid}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link[etl]{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
etl_transform.etl_covid
}
\details{
fill in
}
