% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{japan}
\alias{japan}
\title{Japan Covid-19 situation}
\format{
A list.
\describe{
\item{area_df}{

A tibble.
\describe{
\item{publish_date}{2020-01-16}
\item{population}{5250000}
\item{code}{xxxx}
\item{prefecture}{"Hokkaido"}
\item{concept}{"new_conf"}
\item{value}{0}
\item{value_ma}{0}
\item{value_per1m}{0}
\item{value_ma_per1m}{0}
}
}
\item{map_df}{

A data frame. Each row is prefecture map data. 47 rows.
\describe{
\item{SP_ID}{"1"}
\item{jiscode}{"01"}
\item{name}{"Hokkaido"}
\item{population}{5506419}
\item{region}{"Hokkaido"}
\item{geometry}{sf data}
}
}
\item{area_menu}{

A character vector of length 47.
\describe{
\item{1}{"Hokkaido"}
\item{2}{"Aomori"}
\item{3}{"Iwate"}
\item{4}{"Miyagi"}
\item{5}{"Akita"}
}
}
}
}
\source{
https://www3.nhk.or.jp/news/special/coronavirus/data/
}
\usage{
japan
}
\description{
Japan Covid-19 situation
}
\keyword{datasets}
