% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_file.R
\name{name_file}
\alias{name_file}
\title{Define download file name}
\usage{
name_file(input)
}
\arguments{
\item{input}{A Shiny input.}
}
\value{
A character vector of length 1.
}
\description{
Define download file name
}
\examples{
\dontrun{
name_file(input)
}
}
