% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm helpers.R
\name{panel_reg}
\alias{panel_reg}
\title{Run panel regressios with fixed effects}
\usage{
panel_reg(formula.cur, df, trendwith = "country")
}
\arguments{
\item{formula.cur}{a valid formular}

\item{df}{a data frame}

\item{trendwith}{factor to be interacted with time trend. standard is "country". Viable alternative is "region".}
}
\value{
output of stargazer
}
\description{
Run panel regressios with fixed effects
}
