% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHE-daily.R
\name{phe.daily}
\alias{phe.daily}
\title{Refresh PHE data & plots}
\usage{
phe.daily(
  highlight = c("Hertfordshire", "Norfolk"),
  hcols = c("black", "red3"),
  archive = F
)
}
\arguments{
\item{highlight}{Vector of regions to highlight. Default is c("Hertfordshire", "Norfolk")}

\item{hcols}{Vector of colours to use for highlighted regions. Default is c("black", "orange")}

\item{archive}{Boolean: copy data to hard drive? Default is F}
}
\description{
Download latest data from PHE & produce pdf of standard plots
}
