% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONS.R
\name{read.ons}
\alias{read.ons}
\title{Generic function: read ONS data}
\usage{
read.ons(fnm, snm, cols, header.row, data.rows, rotate = F)
}
\arguments{
\item{fnm}{File name to load from}

\item{snm}{Sheet name or index}

\item{cols}{Columns to read}

\item{header.row}{Row containing dates}

\item{data.rows}{Rows to read}
}
\description{
Generic function: read ONS data
}
