% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time series plots.R
\name{show_measures_of}
\alias{show_measures_of}
\title{Plot timeline of all measures of a country/region}
\usage{
show_measures_of(units = NULL, measures = world.measures, df = world)
}
\arguments{
\item{units}{name of country or region (also takes vector of names)}
}
\value{
ggplot
}
\description{
Plot timeline of all measures of a country/region
}
\examples{
show_measures_of(units="Ger")
show_measures_of(units=c("Iran","Spain","Italy","Germ"))
#show_measures_of(df=world \%>\% filter(in_country=="Mainland China"))
}
