% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time series plots.R
\name{show_measures_of}
\alias{show_measures_of}
\title{Plot timeline of all measures of a country/region}
\usage{
show_measures_of(
  units = world \%>\% filter(pos.new > 1000 & is.na(in_country)) \%>\% pull(label),
  measures = world.measures,
  df = world
)
}
\arguments{
\item{units}{name of country or region (also takes vector of names). As standard shows only unit lables that are countries and had more than 1000 new cases at some point.}

\item{measures}{vector of measures. standard are all in world.measures}

\item{df}{data frame. standard is world.}
}
\value{
ggplot
}
\description{
Plot timeline of all measures of a country/region
}
\examples{
show_measures_of()
show_measures_of(units="Ger")
show_measures_of(units=c("Iran","Spain","Italy","Germ"))
show_measures_of(units=NULL, df=world \%>\% filter(in_country=="Mainland China"))
}
