% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time series plots.R
\name{show_outcome}
\alias{show_outcome}
\title{show outcome as time series including measures taken}
\usage{
show_outcome(
  select_region = "Lombardy",
  select_country = NULL,
  outcome = "pos.new",
  df = NULL,
  total = F,
  rate = F,
  smooth = FALSE
)
}
\arguments{
\item{select_region}{region to show}

\item{select_country}{country to show}

\item{outcome}{character variable describing outcome in all data frame}

\item{df}{alternative data frame to all that contains outcome}

\item{total}{is the total number given and should be transformed to change (standard: TRUE)}

\item{rate}{should the growth rate be computed and shown}

\item{smooth}{should a geom_smooth be added with span=smooth (standard f,smooth=.2 recommended)}
}
\value{

}
\description{
show outcome as time series including measures taken
}
\examples{
show_outcome()
show_outcome(outcome="dead.new",select_country="Germany")
show_outcome(outcome="dead.total",select_country="Germany")
show_outcome(outcome="dead_it")
show_outcome(select_country="Italy",outcome="positive",smooth=.2)
}
