% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa}
\alias{usa}
\title{USA Covid-19 situation}
\format{
A list.
\describe{
\item{area_df}{

A tibble.
\describe{
\item{publish_date}{2020-01-22}
\item{state}{"Alaska"}
\item{population}{731545}
\item{concept}{"cum_conf"}
\item{value}{0}
\item{value_ma}{0}
\item{value_per1m}{0}
\item{value_ma_per1m}{0}
}
}
\item{map_df}{

A tibble.
\describe{
\item{long}{-87.5}
\item{lat}{30.4}
\item{group}{1}
\item{order}{1}
\item{region}{"alabama"}
\item{subregion}{NA_character_}
}
}
\item{area_menu}{

A character vector of length 52. 50 states plus DC and Total.
\describe{
\item{1}{"Total"}
\item{2}{"Alaska"}
\item{3}{"Alabama"}
\item{4}{"Arkansas"}
\item{5}{"Arizona"}
}
}
}
}
\source{
https://usafacts.org/visualizations/coronavirus-covid-19-spread-map/
}
\usage{
usa
}
\description{
USA Covid-19 situation
}
\keyword{datasets}
